requirements.txt
aiocache==0.11.1
aiocqhttp==1.3.0
aiofiles==0.6.0
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
blinker==1.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
croniter==0.3.37
devtools==0.6.1
fastapi==0.63.0
graia-application-mirai==0.15.0
graia-broadcast==0.6.4
h11==0.9.0
h2==4.0.0
hpack==4.0.0
httpcore==0.12.3
httpx==0.16.1
Hypercorn==0.11.2
hyperframe==6.0.0
idna==2.10
iterwrapper==0.1.4
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.2
loguru==0.5.3
lxml==4.6.2
MarkupSafe==1.1.1
msgpack==1.0.2
multidict==5.1.0
natsort==7.1.1
nonebot==1.8.2
Pillow==8.1.0
priority==1.3.0
pydantic==1.7.3
pygtrie==2.4.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
Quart==0.11.5
regex==2020.11.13
requests==2.25.1
rfc3986==1.4.0
services==0.1.1
singledispatchmethod==1.0
six==1.15.0
sniffio==1.2.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.11.8
websockets==8.1
Werkzeug==1.0.1
win32-setctime==1.0.3
wsproto==1.0.0
yarl==1.6.3
